fix id to int

+++ lib/data/web3/web3_utils.dart
@@ -68,7 +68,7 @@ class Web3Utils {
             EthereumAddress.fromHex(address), BigInt.from(i));
         final uri = await nft.tokenURI(nftId);
         final nftParam = {
-          'id': nftId,
+          'id': nftId.toInt(),
           'contract': contract,
           'uri': uri,
         };
